ISSUE-883: add more executor plans for shuffle
