+++ fusequery/query/src/shuffle/mod.rs
@@ -5,7 +5,21 @@
 #[cfg(test)]
 mod plan_scheduler_test;
 
+mod plan_executor;
+mod plan_executor_default;
+mod plan_executor_empty;
+mod plan_executor_local_read_source;
+mod plan_executor_read_source;
+mod plan_executor_remote;
+mod plan_executor_remote_read_source;
 mod plan_scheduler;
 
+pub use plan_executor::ExecutorPlan;
+pub use plan_executor_default::DefaultExecutorPlan;
+pub use plan_executor_empty::EmptyExecutorPlan;
+pub use plan_executor_local_read_source::LocalReadSourceExecutorPlan;
+pub use plan_executor_read_source::ReadSourceExecutorPlan;
+pub use plan_executor_remote::RemoteExecutorPlan;
+pub use plan_executor_remote_read_source::RemoteReadSourceExecutorPlan;
 pub use plan_scheduler::PlanScheduler;
 pub use plan_scheduler::ScheduledActions;
